requirements.txt
asgiref==3.5.0
attrs==19.3.0
Automat==0.8.0
backports.zoneinfo==0.2.1
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
defusedxml==0.7.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.3
django-allauth==0.50.0
django-cors-headers==3.11.0
djangorestframework==3.13.1
entrypoints==0.3
fail2ban==0.11.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
incremental==16.10.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
netifaces==0.10.4
oauthlib==3.1.0
pexpect==4.6.0
psycopg2==2.9.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyGObject==3.36.0
PyHamcrest==1.9.0
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==4.0.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
python-apt==2.0.0+ubuntu0.20.4.7
python-debian===0.1.36ubuntu1
python3-openid==3.2.0
pytz==2022.1
PyYAML==5.3.1
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.2
sqlparse==0.4.2
ssh-import-id==5.10
systemd-python==234
Twisted==18.9.0
ubuntu-advantage-tools==27.6
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
zope.interface==4.7.1
